Copy in remaining Leray Projector code
This update includes a sloppy hack to the remaining Leray projector
code.   For some reason the driver wasn't updated, but code is included
there to initialize U0 with the curl (on the sphere) of a scalar
function, which is divergence-free.  In the previous incarnation, I had
initialized to the gradient of a function, which is NOT curl-free (wow,
Dave).

An overflow bug was found in that the first point generated by ME is [0
0 1], which is singular in dj.  Refer to the notes in the driver for
details (next upload).