Finish Leray Projector rewrite
I know I said the Leray Projector rewrite was completed in a previous
commit, but the recovery code was bad.  This recovery code is still bad,
but at least it returns matrices of the right dimension and avoids that
weird NaN problem.

I think it might be fruitful to investigate the results over many
different curl(u) to see what might be going wrong.