Remove superfluous twoeps2
This is going to take a while to debug, but it seems I left in a few
superfluous factors of twoeps2 in PSI2 in heateqn.m.  The output of PSI
agrees with Mathematica now.  In order to continue debugging, I'll check
for vector orientation, then (in the case of failure) rewrite the code
in Mathematica to cross-check.