Update matrixRBF2.nb with cross-validation code
I've validated heateqn.m up through the definition of A2.  All of this
appears to be numerically correct.  Caveat:  As soon as the points
depart from the sphere, ex and dx break catastrophically.  This is
because they are a basis for vector fields ON THE SURFACE OF THE SPHERE.
A different basis will obviously need to be chosen for different
manifolds.

Moreover, this ex/dx trick strikes me as a little bit artificial--sure,
it "fixes" the system, but it does so only by introducing singular
coordinates.  This sucks!