add dependency; update nsdriver and navierstokes
Added a test for the surface Laplacian in the nsdriver.  Also added
binom.m, which is a more efficient way of computing the binomial
coefficients (courtesy of Grady) used in dsph.m (again, courtesy of
Grady).

getDivFree for some reason produces vector fields that are "better"
(produce much smaller error against the surface Laplacian) than the old
way (veclapdriver.m).  Chalking it up to an error in the latter; more
tests should be conducted at some point.