Update test suite; check suitability of projection
Updated the test suite to include three more tests of the covariant
derivative.  I believe it checks out.

Also checked where it is appropriate to perform the Leray projection
within the timestepping.  Here are things that appear to increase the
rate of blowup (which may or may not coincide with suitability):

- Absent projection within each RK stage
- Absent projection after synthesizing RK stages (U = U + (1/6)...)

This check also demonstrated that AT THIS STAGE, the Leray projector is
idempotent.