Prepare for Ganesh test
Implemented Coriolis force and the Ganesh test in navierstokes.m.
Included a visual check in nsdriver.m of the given solution for the
Ganesh test1.

There appears to be a discrepancy between the RBF solution and the known
solution, particularly when t is small.  I intend to debug by making
sure that the Ganesh forcing is correct.