Repairing Ganesh forcing
Fixed a few sign errors in the Ganesh forcing function, made cancelling
(i.e., superfluous) sign changes elsewhere for consistency.  Verified
that the hand-written Laplacian of the Ganesh function corresponds to
the numerical Laplacian of the same.

Repaired the time argument to the ganesh forcing along RK stages.

Further work will continue to test the forcing term, continuing onward
to verify the behavior of the analytic solution.